Rework GridTestScene
