+++ osu.Framework.Tests/Visual/Drawables/TestSuiteTransformSequence.cs
@@ -13,16 +13,8 @@ using osuTK.Graphics;
 
 namespace osu.Framework.Tests.Visual.Drawables
 {
-    public class TestSuiteTransformSequence : GridTestSuite
+    public class TestSuiteTransformSequence : TestSuite<TestSceneTransformSequence>
     {
-        private readonly Container[] boxes;
-
-        public TestSuiteTransformSequence()
-            : base(4, 3)
-        {
-            boxes = new Container[Rows * Cols];
-        }
-
         protected override void LoadComplete()
         {
             base.LoadComplete();
@@ -41,7 +33,7 @@ namespace osu.Framework.Tests.Visual.Drawables
 
             AddStep($"{nameof(FinishTransforms)}", delegate
             {
-                foreach (var box in boxes)
+                foreach (var box in TestScene.Boxes)
                     box.FinishTransforms();
             });
 
@@ -58,7 +50,7 @@ namespace osu.Framework.Tests.Visual.Drawables
 
             AddStep($"{nameof(ClearTransforms)}", delegate
             {
-                foreach (var box in boxes)
+                foreach (var box in TestScene.Boxes)
                     box.ClearTransforms();
             });
 
@@ -85,16 +77,16 @@ namespace osu.Framework.Tests.Visual.Drawables
                 "Red after 1 sec transform 1 sec in the past",
             };
 
-            for (int i = 0; i < Rows * Cols; ++i)
+            for (int i = 0; i < TestScene.Rows * TestScene.Cols; ++i)
             {
-                Cell(i).Children = new Drawable[]
+                TestScene.Cell(i).Children = new Drawable[]
                 {
                     new SpriteText
                     {
                         Text = labels[i],
                         Font = new FontUsage(size: 20),
                     },
-                    boxes[i] = new Container
+                    TestScene.Boxes[i] = new Container
                     {
                         RelativeSizeAxes = Axes.Both,
                         Size = new Vector2(0.25f),
@@ -120,68 +112,68 @@ namespace osu.Framework.Tests.Visual.Drawables
 
         private void animate()
         {
-            boxes[0].Delay(500).Then(500).Then(500).Then(
+            TestScene.Boxes[0].Delay(500).Then(500).Then(500).Then(
                 b => b.Delay(500).Spin(1000, RotationDirection.CounterClockwise)
             );
 
-            boxes[1].Spin(1000, RotationDirection.CounterClockwise);
+            TestScene.Boxes[1].Spin(1000, RotationDirection.CounterClockwise);
 
-            boxes[2].Delay(-2000).Spin(1000, RotationDirection.CounterClockwise);
+            TestScene.Boxes[2].Delay(-2000).Spin(1000, RotationDirection.CounterClockwise);
 
-            boxes[3].RotateTo(90)
-                    .Then().Delay(1000).RotateTo(0)
-                    .Then().RotateTo(180, 1000).Loop();
+            TestScene.Boxes[3].RotateTo(90)
+                     .Then().Delay(1000).RotateTo(0)
+                     .Then().RotateTo(180, 1000).Loop();
 
-            boxes[4].Delay(1000).Loop(1000, 10, b => b.RotateTo(0).RotateTo(340, 1000));
+            TestScene.Boxes[4].Delay(1000).Loop(1000, 10, b => b.RotateTo(0).RotateTo(340, 1000));
 
-            boxes[5].RotateTo(0).ScaleTo(1).RotateTo(360, 1000)
-                    .Then(1000,
-                        b => b.RotateTo(0, 1000),
-                        b => b.ScaleTo(2, 500)
-                    )
-                    .Then().RotateTo(360, 1000).ScaleTo(0.5f, 1000)
-                    .Then().FadeEdgeEffectTo(Color4.Red, 1000).ScaleTo(2, 500);
+            TestScene.Boxes[5].RotateTo(0).ScaleTo(1).RotateTo(360, 1000)
+                     .Then(1000,
+                         b => b.RotateTo(0, 1000),
+                         b => b.ScaleTo(2, 500)
+                     )
+                     .Then().RotateTo(360, 1000).ScaleTo(0.5f, 1000)
+                     .Then().FadeEdgeEffectTo(Color4.Red, 1000).ScaleTo(2, 500);
 
-            boxes[6].RotateTo(0).ScaleTo(1).RotateTo(360, 500)
-                    .Then(1000,
-                        b => b.RotateTo(0),
-                        b => b.ScaleTo(2)
-                    )
-                    .Then(
-                        b => b.Loop(500, 2, d => d.RotateTo(0).RotateTo(360, 1000)).Delay(500).ScaleTo(0.5f, 500)
-                    )
-                    .Then().FadeEdgeEffectTo(Color4.Red, 1000).ScaleTo(2, 500)
-                    .Finally(_ => finalizeTriggered = true);
+            TestScene.Boxes[6].RotateTo(0).ScaleTo(1).RotateTo(360, 500)
+                     .Then(1000,
+                         b => b.RotateTo(0),
+                         b => b.ScaleTo(2)
+                     )
+                     .Then(
+                         b => b.Loop(500, 2, d => d.RotateTo(0).RotateTo(360, 1000)).Delay(500).ScaleTo(0.5f, 500)
+                     )
+                     .Then().FadeEdgeEffectTo(Color4.Red, 1000).ScaleTo(2, 500)
+                     .Finally(_ => finalizeTriggered = true);
 
-            boxes[7].RotateTo(0).ScaleTo(1).RotateTo(360, 500)
-                    .Then(1000,
-                        b => b.RotateTo(0),
-                        b => b.ScaleTo(2)
-                    )
-                    .Then(
-                        b => b.Loop(500, 2, d => d.RotateTo(0).RotateTo(360, 1000)),
-                        b => b.ScaleTo(0.5f, 500)
-                    )
-                    .OnAbort(b => b.FadeEdgeEffectTo(Color4.Red, 1000));
+            TestScene.Boxes[7].RotateTo(0).ScaleTo(1).RotateTo(360, 500)
+                     .Then(1000,
+                         b => b.RotateTo(0),
+                         b => b.ScaleTo(2)
+                     )
+                     .Then(
+                         b => b.Loop(500, 2, d => d.RotateTo(0).RotateTo(360, 1000)),
+                         b => b.ScaleTo(0.5f, 500)
+                     )
+                     .OnAbort(b => b.FadeEdgeEffectTo(Color4.Red, 1000));
 
-            boxes[8].RotateTo(0).ScaleTo(1).RotateTo(360, 500)
-                    .Then(1000,
-                        b => b.RotateTo(0),
-                        b => b.ScaleTo(2)
-                    )
-                    .Then(
-                        b => b.Loop(500, 2, d => d.RotateTo(0).RotateTo(360, 1000)),
-                        b => b.ScaleTo(0.5f, 500)
-                    )
-                    .Finally(b => b.FadeEdgeEffectTo(Color4.Red, 1000));
-
-            boxes[9].RotateTo(200)
-                    .Finally(b => b.FadeEdgeEffectTo(Color4.Red, 1000));
-
-            boxes[10].Delay(-1000).RotateTo(200)
+            TestScene.Boxes[8].RotateTo(0).ScaleTo(1).RotateTo(360, 500)
+                     .Then(1000,
+                         b => b.RotateTo(0),
+                         b => b.ScaleTo(2)
+                     )
+                     .Then(
+                         b => b.Loop(500, 2, d => d.RotateTo(0).RotateTo(360, 1000)),
+                         b => b.ScaleTo(0.5f, 500)
+                     )
                      .Finally(b => b.FadeEdgeEffectTo(Color4.Red, 1000));
 
-            boxes[11].Delay(-1000).RotateTo(200, 1000)
+            TestScene.Boxes[9].RotateTo(200)
+                     .Finally(b => b.FadeEdgeEffectTo(Color4.Red, 1000));
+
+            TestScene.Boxes[10].Delay(-1000).RotateTo(200)
+                     .Finally(b => b.FadeEdgeEffectTo(Color4.Red, 1000));
+
+            TestScene.Boxes[11].Delay(-1000).RotateTo(200, 1000)
                      .Finally(b => b.FadeEdgeEffectTo(Color4.Red, 1000));
         }
     }
